Separate iiif utility functions into those for dealing with v2 and those for v3
So it’s easier to see what is going on and still needs updating

# Conflicts:
#	catalogue/webapp/pages/api/work.ts
