+++ catalogue/webapp/components/MediaAnnotations/MediaAnnotations.tsx
@@ -1,6 +1,6 @@
 import { ReactElement, FunctionComponent } from 'react';
 import { IIIFMediaElement, IIIFAnnotationResource } from '../../model/iiif';
-import { getAnnotationFromMediaElement } from '../../utils/iiif';
+import { getAnnotationFromMediaElement } from '../../utils/iiif/v2';
 import Space from '@weco/common/views/components/styled/Space';
 import DownloadLink from '@weco/catalogue/components/DownloadLink/DownloadLink';
